refactor: align analog simulation rendering with conventions
