+++ package.json
@@ -50,8 +50,8 @@
     "@tscircuit/matchpack": "^0.0.43",
     "@tscircuit/math-utils": "^0.0.36",
     "@tscircuit/miniflex": "^0.0.4",
-    "@tscircuit/ngspice-spice-engine": "github:tscircuit/ngspice-spice-engine#b2884c969d374b18421f8dd995ca2a4bc6b05229",
-    "@tscircuit/props": "github:tscircuit/props#d2937433997c3be92724d3dbb447398208916693",
+    "@tscircuit/ngspice-spice-engine": "github:tscircuit/ngspice-spice-engine#7bb2401cc8a5ad354a397d293f1f48567a43ed85",
+    "@tscircuit/props": "github:tscircuit/props#a4cb4140629c52ce1870fe9d6e0d48c8fe738c0c",
     "@tscircuit/schematic-match-adapt": "^0.0.18",
     "@tscircuit/schematic-trace-solver": "^0.0.106",
     "@tscircuit/solver-utils": "^0.0.16",
@@ -65,11 +65,12 @@
     "bun-match-svg": "0.0.12",
     "calculate-elbow": "^0.0.12",
     "chokidar-cli": "^3.0.0",
-    "circuit-json": "github:tscircuit/circuit-json#72b4fd2b88decced344cdf8c28d6d36a73696b1b",
+    "circuit-json": "github:tscircuit/circuit-json#2acd7b3d7944bac51afcbc2a7aabc70f712c7680",
     "circuit-json-to-bpc": "^0.0.13",
     "circuit-json-to-connectivity-map": "^0.0.23",
     "circuit-json-to-gltf": "^0.0.107",
-    "circuit-to-svg": "github:tscircuit/circuit-to-svg#8fc8eb5d18807353cd3b67103081d48d0d87bb26",
+    "circuit-json-to-spice": "github:tscircuit/circuit-json-to-spice#14f20d960954c843cd2b020352761429eb8b3638",
+    "circuit-to-svg": "github:tscircuit/circuit-to-svg#d3c30921ec2bbbc76d82f11f972b43c5335acdfe",
     "concurrently": "^9.1.2",
     "connectivity-map": "^1.0.0",
     "debug": "^4.3.6",
@@ -87,6 +88,7 @@
     "react-dom": "^19.1.0",
     "schematic-symbols": "^0.0.233",
     "spicets": "^0.0.2",
+    "spicey": "github:tscircuit/spicey#734c27d20a77e992a7273c5e349f1a1cdada0a72",
     "stack-svgs": "^0.0.1",
     "ts-expect": "^1.3.0",
     "tsup": "^8.2.4"
@@ -114,26 +116,13 @@
     "@lume/kiwi": "^0.4.3",
     "calculate-cell-boundaries": "^0.0.13",
     "calculate-packing": "^0.0.79",
-    "circuit-json-to-spice": "github:tscircuit/circuit-json-to-spice#4eb62507b177c0965d74e08c6800162159adad17",
     "css-select": "5.1.0",
     "format-si-unit": "^0.0.7",
     "nanoid": "^5.0.7",
     "performance-now": "^2.1.0",
     "react-reconciler": "^0.32.0",
-    "spicey": "github:tscircuit/spicey#e969943f6399346297b27ad95ed3b5a6a8845859",
     "svg-path-commander": "^2.1.11",
     "transformation-matrix": "^2.16.1",
     "zod": "^3.25.67"
-  },
-  "trustedDependencies": [
-    "@resvg/resvg-js",
-    "@tscircuit/ngspice-spice-engine",
-    "@tscircuit/props",
-    "circuit-json",
-    "circuit-json-to-spice",
-    "circuit-to-svg",
-    "esbuild",
-    "sharp",
-    "spicey"
-  ]
+  }
 }
